+++ proj4rs/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## Unreleased
 
+* Enforce multi-thread support on native build.
+* Add 'aeqd' projection in default wasm build
 * Implement GeoTIFF grids support
 * Fix deprecated field use in `geo_types::Triangle`
 * Add `krovak`,`cea` and `mill` in default build.